# backend/document/openApi.yaml
openapi: 3.0.3
info:
  title: Playroute API
  version: 1.0.0
  description: PlayrouteのバックエンドAPI。ユーザー・スポット・プラン・プランスポットのCRUDを提供。
servers:
  - url: http://localhost:8080/api
    description: ローカル開発用サーバ
paths:
...
  /users:
    get:
      summary: ユーザー一覧取得
      responses:
        "200":
          description: ユーザー一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      summary: ユーザー作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: 作成完了
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      summary: ユーザー詳細取得
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: ユーザー詳細
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      summary: ユーザー更新
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: 更新完了
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      summary: ユーザー削除
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
      responses:
        "204":
          description: 削除完了

  # =======================
  # スポット
  # =======================
...
components:
  schemas:
    User:
      type: object
      properties:
        user_id:
          type: integer
        user_name:
          type: string
        budget_limit:
          type: number
          format: float
        default_transport:
          type: string
        interests:
          type: string
        created_at:
          type: string
          format: date-time
        created_by:
          type: string
        updated_at:
          type: string
          format: date-time
        updated_by:
          type: string
    UserCreate:
      type: object
      required: [user_name]
      properties:
        user_name:
          type: string
        budget_limit:
          type: number
          format: float
        default_transport:
          type: string
        interests:
          type: string
        created_by:
          type: string
        updated_by:
          type: string
    UserUpdate:
      allOf:
        - $ref: "#/components/schemas/UserCreate"
...